[ci, trainer] feat: add fsdpturbo backend engine support.
