+++ .github/workflows/nightly_ascend.yml
@@ -359,7 +359,7 @@ jobs:
     runs-on: linux-aarch64-a2b3-8
     timeout-minutes: 180 # Increase this timeout value as needed
     container:
-      image: swr.ap-southeast-1.myhuaweicloud.com/base_image/ascend-ci/verl/verl:latest-cann9.0.0-torch_npu2.9.0post2-910b-ubuntu22.04-py3.11-vllm
+      image: swr.cn-southwest-2.myhuaweicloud.com/modelfoundry/ascend-ci/verl/verl:latest-vllm-910b-ubuntu
       options: >-
         --shm-size 16g
     env:
@@ -383,7 +383,7 @@ jobs:
           pip install --no-deps -e .
       - name: Clone FSDPTurbo
         run: |
-          git clone https://gitcode.com/pengnuoheng/FSDPTurbo.git /FSDPTurbo
+          git clone https://gitcode.com/Ascend/FSDPTurbo.git /FSDPTurbo
       - name: Check final pip list
         run: |
           pip list
