+++ verl/workers/engine/fsdp/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .transformer_impl import FSDPEngine, FSDPEngineWithLMHead, FSDPTurboEngineWithLMHead
+from .fsdp_turbo_impl import FSDPTurboEngineWithLMHead
+from .transformer_impl import FSDPEngine, FSDPEngineWithLMHead
 
 __all__ = ["FSDPEngine", "FSDPEngineWithLMHead", "FSDPTurboEngineWithLMHead"]
