+++ verl/workers/engine/fsdp/fsdp_turbo_impl.py
@@ -1,0 +1,151 @@
+# Copyright (c) 2026 Huawei Technologies Co., Ltd. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from contextlib import contextmanager
+
+import torch
+
+from verl.utils.fsdp_utils import fsdp2_load_full_state_dict
+
+from ..base import EngineRegistry
+from .transformer_impl import FSDPEngineWithLMHead
+
+
+@EngineRegistry.register(model_type="language_model", backend="fsdp_turbo", device=["cuda", "npu"])
+class FSDPTurboEngineWithLMHead(FSDPEngineWithLMHead):
+    def _init_device_mesh(self):
+        super()._init_device_mesh()
+        self._init_parallel_state()
+
+    def _init_parallel_state(self):
+        from fsdp_turbo.distributed.parallel_state import get_parallel_state, init_parallel_state
+        from fsdp_turbo.fsdp_turbo_config import FSDPTurboConfig, _dict_to_dataclass
+
+        self.fsdp_turbo_config = _dict_to_dataclass(FSDPTurboConfig, self.engine_config.turbo_config)
+        self.fsdp_turbo_config.distributed.fsdp_plan.cpu_offload = self.engine_config.offload_policy
+        init_parallel_state(self.fsdp_turbo_config)
+        self._parallel_state = get_parallel_state()
+        if self._is_ulysses_enabled():
+            self._process_ulysses_config()
+
+    def _build_module(self):
+        # Do not let verl's Qwen VLM monkey patch slice the text model before
+        # FSDP-Turbo's post-fusion model patch does the CP split.
+        cp_size = self.ulysses_sequence_parallel_size
+        self.ulysses_sequence_parallel_size = 1
+        try:
+            return super()._build_module()
+        finally:
+            self.ulysses_sequence_parallel_size = cp_size
+
+    def _build_fsdp_module(self, module):
+        from fsdp_turbo.fsdp_turbo import FSDPTurbo
+
+        full_state = module.state_dict()
+        module = FSDPTurbo(self.fsdp_turbo_config, module).model
+        offload_policy = None
+        if self.engine_config.offload_policy or self.engine_config.forward_only:
+            self._is_offload_param = False
+            self._is_offload_optimizer = False
+            offload_policy = True
+            self._uses_fsdp2_cpu_offload_policy = True
+        fsdp2_load_full_state_dict(module, full_state, None, offload_policy)
+
+        return module
+
+    def _is_ulysses_enabled(self):
+        return self._parallel_state.is_group_enable("ulysses")
+
+    def _process_ulysses_config(self):
+        if self.ulysses_sequence_parallel_size > 1:
+            raise ValueError(
+                "Do not enable both FSDP-Turbo CP and verl Ulysses SP. "
+                "Use fsdp_kwargs.distributed.ulysses_parallel_size for Turbo CP "
+                "and set ulysses_sequence_parallel_size=1."
+            )
+
+        self.model_config.hf_config._attn_implementation = "eager"
+        self.ulysses_sequence_parallel_size = self._parallel_state.get_ulysses_group_size()
+        self.ulysses_parallel_group = self._parallel_state.get_ulysses_group()
+        self.use_ulysses_sp = True
+
+    def get_data_parallel_rank(self):
+        if not hasattr(self, "_parallel_state"):
+            return super().get_data_parallel_rank()
+        return self._parallel_state.get_data_rank()
+
+    def get_data_parallel_size(self):
+        if not hasattr(self, "_parallel_state"):
+            return super().get_data_parallel_size()
+        return self._parallel_state.get_data_group_size()
+
+    def get_data_parallel_group(self):
+        if not hasattr(self, "_parallel_state"):
+            return super().get_data_parallel_group()
+        return self._parallel_state.get_data_group()
+
+    def is_mp_src_rank_with_outputs(self):
+        if not hasattr(self, "_parallel_state"):
+            return super().is_mp_src_rank_with_outputs()
+        if self._is_ulysses_enabled():
+            is_collect = self._parallel_state.get_ulysses_rank() == 0
+        else:
+            is_collect = True
+        return is_collect
+
+    @contextmanager
+    def _gradient_sync_context(self, *, is_last_micro_batch: bool):
+        # To avoid OOM for fsdp_turbo backend
+        yield
+
+    def optimizer_step(self):
+        """
+        Clip gradients, skip update if non-finite, and step optimizer.
+
+        Returns:
+            grad_norm (float): Norm of gradients before clipping.
+        """
+        assert self.optimizer_config.clip_grad is not None
+
+        # getattr fallback: some subclasses (e.g. VeOmniEngine) bypass FSDPEngine.__init__.
+        scaler = getattr(self, "scaler", None)
+
+        # Unscale gradients before clip so the clip threshold is applied to true gradient
+        # magnitudes, not scaled ones. scaler.step() will skip the update if any grad is inf/nan.
+        if scaler is not None:
+            scaler.unscale_(self.optimizer)
+
+        from fsdp_turbo.training.clip_grads import clip_grad_norm
+
+        grad_norm_value = clip_grad_norm(model=self.module, max_norm=self.optimizer_config.clip_grad)
+        grad_norm = torch.tensor(grad_norm_value, device=next(self.module.parameters()).device, dtype=torch.float32)
+
+        if scaler is not None:
+            # scaler handles inf/nan skipping internally via _check_inf_per_device.
+            scaler.step(self.optimizer)
+            scaler.update()
+        else:
+            # if grad_norm is not finite, skip the update
+            if not torch.isfinite(grad_norm):
+                print(f"WARN: grad_norm is not finite: {grad_norm}")
+                self.optimizer.zero_grad()
+            else:
+                self.optimizer.step()
+
+        if self._qat_enabled:
+            from verl.utils.qat.core import invalidate_all_scales
+
+            invalidate_all_scales(self.module)
+
+        return grad_norm.item()
